создание и изменение проекта

+++ app/src/main/java/ru/tinkoff/tinkoffer/data/models/projects/response/ProjectInfoDto.kt
@@ -8,8 +8,8 @@ data class ProjectInfoDto(
     val id: String,
     val name: String,
     val schedule: String,
-    val votesPerPeriod: Long,
-    val votesRefreshPeriodDays: Long,
+    val votesPerPeriod: Int,
+    val votesRefreshPeriodDays: Int,
     val createdAt: String,
     val users: List<ProjectUserInfoDto>,
 )
